chore: enable research records
